Ensure DynamicMap.clone Callable references original DynamicMap
